requirements.txt
asgiref==3.6.0
attrs==22.2.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
distro==1.8.0
Django==4.1.7
django-cors-headers==3.13.0
django-filter==22.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
docker==6.0.1
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
drf-extensions==0.7.1
gunicorn==20.1.0
idna==3.4
jsonschema==3.2.0
packaging==23.0
paramiko==3.0.0
Pillow==9.4.0
psycopg2-binary==2.9.5
pycparser==2.21
PyNaCl==1.5.0
pyrsistent==0.19.3
python-decouple==3.7
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==5.4.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
stripe==5.2.0
texttable==1.6.7
urllib3==1.26.14
websocket-client==0.59.0
